% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_methods.R
\name{coef_names}
\alias{coef_names}
\title{List Available Coefficients}
\usage{
coef_names(object)
}
\arguments{
\item{object}{A \code{qqfit} object.}
}
\value{
Character vector of coefficient names.
}
\description{
List Available Coefficients
}
