% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{compute_slice_ylim}
\alias{compute_slice_ylim}
\title{Compute Global Y-axis Limits for Slice Plots}
\usage{
compute_slice_ylim(coefs, se, taus, taus_slice, slice_dim = c("u", "v"))
}
\description{
Compute Global Y-axis Limits for Slice Plots
}
\keyword{internal}
