% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_num}
\alias{format_num}
\title{Format Numbers for Display}
\usage{
format_num(x, digits = 4)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Number of decimal places.}
}
\value{
Character vector.
}
\description{
Format Numbers for Display
}
\keyword{internal}
