% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_between_var}
\alias{is_between_var}
\title{Check if Variable is Constant Within Groups}
\usage{
is_between_var(x, group, tol = 1e-10)
}
\arguments{
\item{x}{A vector.}

\item{group}{A group indicator vector.}

\item{tol}{Tolerance for checking variance.}
}
\value{
Logical, TRUE if x is constant within all groups.
}
\description{
Check if Variable is Constant Within Groups
}
\keyword{internal}
