% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_within_var}
\alias{is_within_var}
\title{Check if Variable Varies Within All Groups}
\usage{
is_within_var(x, group, tol = 1e-10)
}
\arguments{
\item{x}{A vector.}

\item{group}{A group indicator vector.}

\item{tol}{Tolerance for checking variance.}
}
\value{
Logical, TRUE if x varies within all groups.
}
\description{
Check if Variable Varies Within All Groups
}
\keyword{internal}
