% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{plot_single}
\alias{plot_single}
\title{Internal: Plot Single Coefficient}
\usage{
plot_single(x, which, type, taus_slice, ncol, ylim, ...)
}
\description{
Internal: Plot Single Coefficient
}
\keyword{internal}
