% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{plot_slices_u}
\alias{plot_slices_u}
\title{Internal: Slice Plot Fixing u}
\usage{
plot_slices_u(
  coefs,
  se,
  taus,
  taus_slice,
  coef_name,
  ncol,
  shared_ylim,
  ylim,
  ylab = NULL,
  xlab = NULL,
  ggtitle = NULL,
  panel_prefix = NULL,
  line_color = "#11333a",
  ribbon_fill = "#0faeb6",
  point_color = NULL,
  ribbon_alpha = 0.2,
  ...
)
}
\description{
Internal: Slice Plot Fixing u
}
\keyword{internal}
