% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{plot_slices_v}
\alias{plot_slices_v}
\title{Internal: Slice Plot Fixing v}
\usage{
plot_slices_v(coefs, se, taus, taus_slice, coef_name, ncol, ylim, ...)
}
\description{
Internal: Slice Plot Fixing v
}
\keyword{internal}
