% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{plot_surface}
\alias{plot_surface}
\title{Internal: Surface Plot}
\usage{
plot_surface(
  coefs,
  se,
  taus,
  coef_name,
  ggtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\description{
Internal: Surface Plot
}
\keyword{internal}
