% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_methods.R
\name{print.qqfit}
\alias{print.qqfit}
\title{Print Method for qqfit Objects}
\usage{
\method{print}{qqfit}(x, ...)
}
\arguments{
\item{x}{A \code{qqfit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for qqfit Objects
}
