% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_bootstrap.R
\name{qq_bootstrap}
\alias{qq_bootstrap}
\title{Clustered Bootstrap for QQ Regression}
\usage{
qq_bootstrap(
  fitted_first,
  X,
  group_vec,
  taus,
  nboot,
  parallel = TRUE,
  ncores = NULL,
  weights = NULL
)
}
\arguments{
\item{fitted_first}{Matrix of first-stage fitted values (n x n_tau).}

\item{X}{Full design matrix.}

\item{group_vec}{Group indicator vector.}

\item{taus}{Numeric vector of quantile levels.}

\item{nboot}{Number of bootstrap replications.}

\item{parallel}{Logical, whether to parallelize.}

\item{ncores}{Number of cores to use if parallel = TRUE.}

\item{weights}{Optional observation weights.}
}
\value{
A list with:
\item{se}{Array of bootstrap standard errors (k x n_tau x n_tau)}
\item{coef_boot}{Array of bootstrap coefficient draws (k x n_tau x n_tau x nboot)}
}
\description{
Performs clustered bootstrap inference by resampling entire groups
with replacement. Parallelization is done over bootstrap iterations
(nboot) since nboot >> n_tau.
}
\keyword{internal}
