% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{qqreg-package}
\alias{qqreg}
\alias{qqreg-package}
\title{qqreg: Quantile on Quantiles Regression}
\description{
Implements the two-step quantile regression estimator from
Pons (2026) "Quantile on Quantiles". The package estimates a
two-dimensional quantile surface capturing heterogeneity both
within and between groups.
}
\section{Main function}{

\describe{
\item{\code{\link{qq_fit}}}{Fit a QQ regression model}
}
}

\section{Methods}{

\describe{
\item{\code{print.qqfit}}{Print model summary}
\item{\code{summary.qqfit}}{Detailed model summary}
\item{\code{coef.qqfit}}{Extract coefficients}
\item{\code{plot.qqfit}}{Plot coefficient surfaces}
}
}

\author{
\strong{Maintainer}: Martina Pons \email{martina.pons@econ.uzh.ch}

}
