% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{plot_slices_u}
\alias{plot_slices_u}
\title{Internal: Slice Plot Fixing u}
\usage{
plot_slices_u(coefs, se, taus, taus_slice, coef_name, ncol, ylim, ...)
}
\description{
Internal: Slice Plot Fixing u
}
\keyword{internal}
